Add Mailboxes
Add procedure to Message_Manager for module's to register themselves.
This procedure supplies them with a Module_Mailbox with several capabilities:
	- The only way to read the module's messages is with one
	of these boxes.
	- The only (expected) way for modules to send messages is
	by using a mailbox. The mailbox will ensure the message is
	sent from the correct address.
Additionally, Messages are now passed around with access types
and mailboxes have a queue size determined per module.

The tests still pass, but the SPARK analysis is incomplete. Waiting
until refactor is done to do that.