Continue message refactor, switch to Alire (#35)
I've done some work on Merc. I've updated the API packages for
Time_Server, File_Server and the Pub/Sub server. They now have 
strict pre-conditions to ensure that messages are only sent to modules
which can receive them. Note that this functionality isn't implemented
yet so the preconditions just fail.

Now that Messages have dynamic payload sizes which are access types,
there are some additional memory management steps when using 
messages. Encoder functions are now procedures because according
to SPARK, allocating memory is a global effect. I have made the syntax
of sending messages bearable by making Merc generate an additional
Send_Message_Type procedure for all messages.

Merc is now able to generate 100% of the Time_Server api. The
pub/sub and file servers are mostly there, and in this commit
I copy/paste hacked the generated Merc bits and formal proofs
into the packages for them.

I've also converted the project to Alire. This will give us access to the SPARK
library, which has formal versions of the standard library containers. CubedOS
currently compiles and does everything it needs to as an Alire project,
but there are some more things to do in that area:

- How to make CubedOS a library. This hasn't happened because libraries
aren't allowed to have executables (and that's how we run the tests). This is
just a matter of reading the instructions on adding testing the right way.
- Make the GPR structure best practice.
- Set up testing